Version 1.1
Add

+++ ProyectoMysql/Api/LogisticStorage/LogisticStorage.EntityLayer/DespachoDetalleEntity.cs
@@ -23,7 +23,7 @@ namespace LogisticStorage.EntityLayer
 
         public Int32 DespachoDetalleId { get; set; }
         public Int32 DespachoId { get; set; }
-        public Int32 Cantidad { get; set; }    
+        public Decimal Cantidad { get; set; }    
         public Int32 OrdenPedidoId { get; set; }
         public Int32 OrdenPedidoDetalleId { get; set; }
         public String NomProducto { get; set; }
